updated expression tables, give scopes and namespaces separate structures

+++ src/structures/string/str.c
@@ -27,7 +27,7 @@ char* str_create(int* len, int* alloc, const char* val) {
 		str = malloc(sizeof(char) * (*alloc + 1));
 		str[*alloc] = '\0'; // null terminator
 	} else {
-		*alloc = strlen(val);
+		*alloc = (int)strlen(val);
 		*len = *alloc - 1;
 		str = malloc(sizeof(char) * (*alloc + 1));
 		strcpy(str, val);
@@ -55,7 +55,7 @@ char* str_add_char(char* str, int* len, int* alloc, char val) {
 char* str_add_str(char* str, int* len, int* alloc, const char* val) {
 	
 	// reallocate if necessary
-	int new_alloc = strlen(val) + *len;
+	int new_alloc = (int)strlen(val) + *len;
 	if (new_alloc >= *alloc) {
 		*alloc = new_alloc;
 		str = realloc(str, sizeof(char) * (*alloc + 1));
@@ -74,7 +74,7 @@ char* str_add_str(char* str, int* len, int* alloc, const char* val) {
 char* str_copy(const char* val, int* len) {
 	/*	it's ok to use an int instead of unsigned long,
 	 	this is for error messages, and this isn't C++ */
-	int val_len = strlen(val);
+	int val_len = (int)strlen(val);
 	char* str = malloc(sizeof(char) * val_len + 1);
 	strcpy(str, val);
